General fixes:   Allow whitespace before ';' in cascades.   Fix nested literal arrays.   Improve messages when a variable used is unknown.   Give metaclasses the name 'ClassName class' (depending on the class's name).   Add ClassDescription>>name.   Provide Object>>-> to create an Association.   Get Transcript to use the printString of non-String objects to produce better output results.
There is a bug with block returns in nested blocks and called methods.
I cannot see an obvious way to fix this, so I have had to code many
methods with top-level answers instead of answering from in a block (see
Dictionary>>findKey:ifAbsent:, Set>>find:ifAbsent:, Set>>occurrencesOf:,
etc).

The following classes have been implemented with tests as far as the
Blue Book specification goes:

Association
Bag
Dictionary
LookupKey

Note that I have implemented Dictionary (and Set and Bag) exactly as
required in the Blue Book and not by using an underlying HashMap or
HashSet.  This is to provide an initial consistent base.  If it is
considered important to switch the underlying implementation, then this
can be done if the tests remain working.